+++ contentstack/__init__.py
@@ -2,7 +2,6 @@
 The __init__.py files are required to make Python treat the directories as containing
 packages; this is done to prevent directories with a common name, such as string,
 from unintentionally hiding valid modules that occur later on the module search path
-
 """
 from .entry import Entry
 from .asset import Asset
@@ -14,7 +13,7 @@ from .utility import Utils
 __title__ = 'contentstack-python'
 __author__ = 'contentstack'
 __status__ = 'debug'
-__version__ = '1.8.0'
+__version__ = 'v1.8.0'
 __endpoint__ = 'cdn.contentstack.io'
 __email__ = 'mobile@contentstack.com'
 __developer_email__ = 'shailesh.mishra@contentstack.com'