+++ tests/test_live_preview.py
@@ -3,13 +3,18 @@ import unittest
 import config
 import contentstack
 
-_preview = {
-    'enable': True,
-    'live_preview': '#*#*#*#*#',
-    'host': 'cdn.contentstack.io',
+management_token = 'cs8743874323343u9'
+entry_uid = 'blt8743874323343u9'
+
+_lp_query = {
+    'live_preview': '#0#0#0#0#0#0#0#0#0#',
     'content_type_uid': 'product',
-    'entry_uid': 'blt909u787843',
-    'authorization': 'management_token@fake@testing'
+    'entry_uid': entry_uid
+}
+_lp = {
+    'enable': True,
+    'host': 'api.contentstack.io',
+    'management_token': management_token
 }
 
 API_KEY = config.APIKEY
@@ -26,59 +31,59 @@ class TestLivePreviewConfig(unittest.TestCase):
             API_KEY, DELIVERY_TOKEN,
             ENVIRONMENT, host=HOST)
 
-    def test_01_live_preview_enabled_(self):
+    def test_live_preview_disabled(self):
         self.stack = contentstack.Stack(
             API_KEY,
             DELIVERY_TOKEN,
             ENVIRONMENT,
-            live_preview=_preview)
-        self.stack.content_type(
-            'live_content_type').entry('live_entry_uid')
+            live_preview={
+                'enable': False,
+                'host': 'api.contentstack.io',
+                'management_token': 'string987654321'
+            })
+        self.stack.content_type('product').entry(entry_uid)
         self.assertEqual(3, len(self.stack.get_live_preview))
-        self.assertTrue(self.stack.get_live_preview['enable'])
-        self.assertTrue(self.stack.get_live_preview['authorization'])
+        self.assertFalse(self.stack.get_live_preview['enable'])
+        self.assertTrue(self.stack.get_live_preview['management_token'])
 
-    def test_021_live_preview_enabled_(self):
+    def test_021_live_preview_enabled(self):
         self.stack = contentstack.Stack(
             API_KEY,
             DELIVERY_TOKEN,
             ENVIRONMENT,
-            live_preview=_preview)
-        self.assertEqual(_preview['authorization'],
-                         self.stack.live_preview['authorization'])
+            live_preview=_lp)
+        self.stack.live_preview_query(live_preview_query=_lp_query)
+        self.assertIsNotNone(self.stack.live_preview['management_token'])
+        self.assertEqual(7, len(self.stack.live_preview))
+        self.assertEqual('product', self.stack.live_preview['content_type_uid'])
 
     def test_03_set_host(self):
         self.stack = contentstack.Stack(
             API_KEY,
             DELIVERY_TOKEN,
             ENVIRONMENT,
-            live_preview=_preview)
-        self.assertEqual(3, len(self.stack.live_preview))
-        self.assertEqual(True, self.stack.live_preview['enable'])
+            live_preview=_lp)
+        self.assertEqual(7, len(self.stack.live_preview))
+        self.assertEqual(True, 'enable' in self.stack.live_preview)
 
     def test_031_set_host_value(self):
         self.stack = contentstack.Stack(
             API_KEY,
             DELIVERY_TOKEN,
             ENVIRONMENT,
-            live_preview=_preview)
-        self.assertEqual(3, len(self.stack.live_preview))
-        self.assertEqual(_preview['host'],
-                         self.stack.live_preview['host'])
+            live_preview=_lp)
+        self.stack.live_preview_query(live_preview_query=_lp_query)
+        self.assertEqual(7, len(self.stack.live_preview))
+        self.assertIsNotNone(self.stack.live_preview['host'])
 
     def test_06_live_preview_query(self):
-        _live_preview = {
-            'include_edit_tags': True,
-            'edit_tags_type': object,
-        }
-        _preview.update(_live_preview)
         self.stack = contentstack.Stack(
             API_KEY,
             DELIVERY_TOKEN,
             ENVIRONMENT,
-            live_preview=_preview
+            live_preview=_lp
         )
-        self.assertEqual(5, len(self.stack.live_preview))
+        self.assertEqual(7, len(self.stack.live_preview))
 
     def test_07_branching(self):
         stack = contentstack.Stack(
@@ -91,11 +96,9 @@ class TestLivePreviewConfig(unittest.TestCase):
             API_KEY,
             DELIVERY_TOKEN,
             ENVIRONMENT,
-            live_preview=_preview
+            live_preview=_lp
         )
-        self.stack.live_preview_query(
-            hash='live_preview',
-            content_type_uid='fake@content_type')
+        self.stack.live_preview_query(live_preview_query=_lp_query)
         self.assertEqual(7, len(self.stack.live_preview))
 
     def test_09_live_preview_query_hash_excluded(self):
@@ -103,10 +106,10 @@ class TestLivePreviewConfig(unittest.TestCase):
             API_KEY,
             DELIVERY_TOKEN,
             ENVIRONMENT,
-            live_preview=_preview
+            live_preview=_lp
         )
-        self.stack.live_preview_query(live_preview=_preview)
-        self.stack.content_type('product').entry(entry_uid='blt43434433432')
+        self.stack.live_preview_query(live_preview_query=_lp_query)
+        self.stack.content_type('product').entry(entry_uid=entry_uid)
         self.assertEqual(3, len(self.stack.headers))
         self.assertEqual(True, 'access_token' in self.stack.headers)
         self.assertEqual(True, 'api_key' in self.stack.headers)
@@ -116,32 +119,16 @@ class TestLivePreviewConfig(unittest.TestCase):
             API_KEY,
             DELIVERY_TOKEN,
             ENVIRONMENT,
-            live_preview=_preview
+            live_preview=_lp
         )
-        self.stack.live_preview_query(live_preview={
-            'enable': True,
-            'live_preview': 'B0B)B)B)BB)B',
-            'host': 'cdn.contentstack.io',
-            'content_type_uid': 'producto',
-            'entry_uid': 'blt909u787843',
-            'authorization': 'management_token@fake@testing'
-        })
+        self.stack.live_preview_query(live_preview_query=_lp_query)
         entry = self.stack.content_type('product').entry(entry_uid=ENTRY_UID)
-        response = entry.fetch()
-        self.assertEqual(2, len(self.stack.headers))
+        resp = entry.fetch()
+        print(resp)
+        self.assertEqual(6, len(self.stack.headers))
         self.assertEqual(API_KEY, self.stack.headers['api_key'])
-
-    def test_11_live_preview(self):
-        stack = contentstack.Stack(
-            'api_key',
-            'delivery_token',
-            'development',
-            live_preview=_preview).live_preview_query(
-            entry_uid='entry_uid',
-            content_type_uid='bugfixes',
-            hash='#Just_fake_it'
-        )
-        result = stack.content_type(content_type_uid='bugfixes').query().find()
+        self.assertEqual(DELIVERY_TOKEN, self.stack.headers['access_token'])
+        self.assertEqual(ENVIRONMENT, self.stack.headers['environment'])
 
 
 if __name__ == '__main__':