+++ CHANGELOG.md
@@ -2,11 +2,12 @@
 
 ## _v1.8.0_
 
-### **Date: 09-MAY-2023**
+### **Date: 23-MAY-2023**
 
-- AZURE_EU, region support added
-- include_metadata support added to asset, entry and query
+- AZURE_EU, Region support added
+- Include Metadata support added to asset, entry and query
 - General code improvement clean up
+- Updated code for Live Preview
 
 ---