+++ changelog.rst
@@ -5,12 +5,12 @@
 *v1.8.0*
 ============
 
-**Date: 10-MAY-2023**
+**Date: 23-MAY-2023**
 
 - AZURE_EU, Region support added
 - Include Metadata support added to asset, entry and query
 - General code improvement clean up
-- Changes done for Live Preview
+- Updated code for Live Preview
 
 
 *v1.7.0*